feat: declare generation arguments to enable silent usage
Declaring the arguments makes sure yo ts-app --help does display and
explain the possible arguments. It also enables the generator to run
without any user interaction, e.g. from within a code generation tool.

The two boolean ones (newdir, initrepo) have been modeled as
"arguments", not as "options" (=flags) because as options the difference
between wanting them to be "false" and not wanting to set them as
commandline arguments could not be recognized.